Separate course wide and user session results
