+++ src/it/java/teammates/it/ui/webapi/GetCourseSessionResultsActionIT.java
@@ -1,0 +1,92 @@
+package teammates.it.ui.webapi;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.Test;
+
+import teammates.common.datatransfer.DataBundle;
+import teammates.common.datatransfer.FeedbackResultFetchType;
+import teammates.common.util.Const;
+import teammates.common.util.HibernateUtil;
+import teammates.storage.entity.Course;
+import teammates.storage.entity.FeedbackSession;
+import teammates.storage.entity.Instructor;
+import teammates.ui.output.SessionResultsData;
+import teammates.ui.request.Intent;
+import teammates.ui.webapi.GetCourseSessionResultsAction;
+import teammates.ui.webapi.JsonResult;
+
+/**
+ * SUT: {@link GetCourseSessionResultsAction}.
+ */
+public class GetCourseSessionResultsActionIT extends BaseActionIT<GetCourseSessionResultsAction> {
+    private DataBundle typicalBundle;
+
+    @Override
+    protected String getActionUri() {
+        return Const.ResourceURIs.COURSE_SESSION_RESULTS;
+    }
+
+    @Override
+    protected String getRequestMethod() {
+        return GET;
+    }
+
+    @Override
+    @BeforeMethod
+    protected void setUp() throws Exception {
+        super.setUp();
+        logoutUser();
+        typicalBundle = persistDataBundle(getTypicalDataBundle());
+        HibernateUtil.flushSession();
+        HibernateUtil.clearSession();
+    }
+
+    @Override
+    @Test
+    protected void testExecute() {
+        Instructor instructor = typicalBundle.instructors.get("instructor1OfCourse1");
+        loginAsInstructor(instructor.getGoogleId());
+
+        FeedbackSession feedbackSession = typicalBundle.feedbackSessions.get("session1InCourse1");
+        String[] params = new String[] {
+                Const.ParamsNames.FEEDBACK_SESSION_ID, feedbackSession.getId().toString(),
+                Const.ParamsNames.INTENT, Intent.FULL_DETAIL.name(),
+        };
+
+        GetCourseSessionResultsAction action = getAction(params);
+        JsonResult result = getJsonResult(action);
+        SessionResultsData output = (SessionResultsData) result.getOutput();
+
+        SessionResultsData expected = SessionResultsData.init(logic.getSessionResults(
+                feedbackSession, instructor.getEmail(), null, null, FeedbackResultFetchType.BOTH));
+
+        assertEquals(expected.getQuestions().size(), output.getQuestions().size());
+    }
+
+    @Override
+    protected void testAccessControl() throws Exception {
+        FeedbackSession feedbackSession = typicalBundle.feedbackSessions.get("session1InCourse1");
+        Course course = typicalBundle.courses.get("course1");
+
+        String[] params = new String[] {
+                Const.ParamsNames.FEEDBACK_SESSION_ID, feedbackSession.getId().toString(),
+                Const.ParamsNames.INTENT, Intent.FULL_DETAIL.name(),
+        };
+
+        verifyAccessibleForInstructorsOfTheSameCourse(course, params);
+        verifyInaccessibleForInstructorsOfOtherCourses(course, params);
+
+        loginAsStudent(typicalBundle.students.get("student1InCourse1").getGoogleId());
+        verifyCannotAccess(params);
+
+        verifyInaccessibleForStudentsOfOtherCourse(course, params);
+
+        String[] invalidIntentParams = new String[] {
+                Const.ParamsNames.FEEDBACK_SESSION_ID, feedbackSession.getId().toString(),
+                Const.ParamsNames.INTENT, Intent.STUDENT_RESULT.name(),
+        };
+        verifyHttpParameterFailure(invalidIntentParams);
+    }
+}
