+++ src/main/java/teammates/ui/constants/ResourceEndpoints.java
@@ -28,7 +28,8 @@ public enum ResourceEndpoints {
     INSTRUCTORS(ResourceURIs.INSTRUCTORS),
     INSTRUCTOR(ResourceURIs.INSTRUCTOR),
     INSTRUCTOR_PRIVILEGE(ResourceURIs.INSTRUCTOR_PRIVILEGE),
-    RESULT(ResourceURIs.RESULT),
+    COURSE_SESSION_RESULTS(ResourceURIs.COURSE_SESSION_RESULTS),
+    USER_SESSION_RESULTS(ResourceURIs.USER_SESSION_RESULTS),
     STUDENTS(ResourceURIs.STUDENTS),
     STUDENT(ResourceURIs.STUDENT),
     USER_KEY(ResourceURIs.USER_KEY),
