+++ src/test/java/teammates/ui/webapi/GetActionClassesActionTest.java
@@ -69,7 +69,8 @@ public class GetActionClassesActionTest extends BaseActionTest<GetActionClassesA
                 GetActionClassesAction.class,
                 UnpublishFeedbackSessionAction.class,
                 PublishFeedbackSessionAction.class,
-                GetSessionResultsAction.class,
+                GetCourseSessionResultsAction.class,
+                GetUserSessionResultsAction.class,
                 GetHasResponsesAction.class,
                 RestoreCourseAction.class,
                 BinCourseAction.class,
