+++ src/web/services/feedback-session-actions.service.ts
@@ -11,7 +11,6 @@ import { SimpleModalType } from '../app/components/simple-modal/simple-modal-typ
 import { ErrorMessageOutput } from '../app/error-message-output';
 import { InstructorSessionResultSectionType } from '../app/pages-instructor/instructor-session-result-page/instructor-session-result-section-type.enum';
 import { FeedbackQuestion } from '../types/api-output';
-import { Intent } from '../types/api-request';
 
 /**
  * Handles sessions related actions.
@@ -33,7 +32,6 @@ export class FeedbackSessionActionsService {
     courseId: string,
     feedbackSessionName: string,
     feedbackSessionId: string,
-    intent: Intent,
     indicateMissingResponses: boolean,
     showStatistics: boolean,
     questions: FeedbackQuestion[],
@@ -60,7 +58,6 @@ export class FeedbackSessionActionsService {
       ...questions.map((question: FeedbackQuestion) =>
         this.feedbackSessionsService.downloadSessionResults(
           feedbackSessionId,
-          intent,
           indicateMissingResponses,
           showStatistics,
           question.feedbackQuestionId,
