+++ src/web/services/feedback-sessions.service.spec.ts
@@ -105,17 +105,28 @@ describe('FeedbackSessionsService', () => {
     expect(spyHttpRequestService.get).toHaveBeenCalledWith(ResourceEndpoints.SESSION_STATS, paramMap);
   });
 
-  it('should call get when retrieving feedback session results', () => {
+  it('should call get when retrieving course feedback session results', () => {
     const paramMap: Record<string, string> = {
       fsid: '248b1915-5f52-4730-b5b2-3ec25a2caabc',
-      intent: Intent.FULL_DETAIL,
     };
 
-    service.getFeedbackSessionResults({
+    service.getCourseSessionResults({
       feedbackSessionId: paramMap['fsid'],
-      intent: Intent.FULL_DETAIL,
     });
-    expect(spyHttpRequestService.get).toHaveBeenCalledWith(ResourceEndpoints.RESULT, paramMap);
+    expect(spyHttpRequestService.get).toHaveBeenCalledWith(ResourceEndpoints.COURSE_SESSION_RESULTS, paramMap);
+  });
+
+  it('should call get when retrieving user feedback session results', () => {
+    const paramMap: Record<string, string> = {
+      fsid: '248b1915-5f52-4730-b5b2-3ec25a2caabc',
+      intent: Intent.STUDENT_RESULT,
+    };
+
+    service.getUserSessionResults({
+      feedbackSessionId: paramMap['fsid'],
+      intent: Intent.STUDENT_RESULT,
+    });
+    expect(spyHttpRequestService.get).toHaveBeenCalledWith(ResourceEndpoints.USER_SESSION_RESULTS, paramMap);
   });
 
   it('should call put when moving session to recycle bin', () => {
