+++ src/web/types/api-const.ts
@@ -77,7 +77,8 @@ export enum ResourceEndpoints {
   INSTRUCTORS = "/webapi/instructors",
   INSTRUCTOR = "/webapi/instructor",
   INSTRUCTOR_PRIVILEGE = "/webapi/instructor/privilege",
-  RESULT = "/webapi/result",
+  COURSE_SESSION_RESULTS = "/webapi/session/results/course",
+  USER_SESSION_RESULTS = "/webapi/session/results/user",
   STUDENTS = "/webapi/students",
   STUDENT = "/webapi/student",
   USER_KEY = "/webapi/user/key",
